Add `data-unique-id-prefix="PREFIX"` attribute recognition in `<svg>` tag.
This attribute allows to specify the prefix string used for all IDs of a given SVG file to avoid duplicates when loading multiple SVGs and is intended to be used instead of the default numeric suffix appended by svg-loader, unless `data-unique-ids="disabled"` is also specified (in which case neither the suffix, nor the prefix are added).

This is especially handy for SVGs containing `<defs><symbol id="...">...</defs>` sections where the prefix can be used to isolate all IDs of a particular SVG with a meaningful name so that these IDs can be later referred to via `<svg><use href="#PREFIX-ORIGINAN_ID"/></svg>` to have the specific symbol rendered on the page.